requirements.txt
streamlit
tensorflow
numpy
Pillow
h5py
protobuf==3.20.0